move the handling of TimeoutException to core.plugins
